git-svn-id: file:///home/nbingham/backup/private-repos/trunk/Software/Work/chp@392 404bc64d-9d1b-498e-bf1b-9a6e5dac48b1

+++ src/data/triple.h
@@ -28,7 +28,7 @@ namespace std
 	template<>
 	struct hash<triple>
 	{
-		inline size_t operator()(const triple &v) const
+		inline int operator()(const triple &v) const
 		{
 			return hash_pair(v.i, hash_pair(v.l, v.h)) % 15485863;
 		}
